+++ index.html
@@ -8,45 +8,33 @@
       rel="stylesheet"
       href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css"
     />
-    <link rel="stylesheet" href="style.css" />
+    <link rel="stylesheet" href="index.css" />
   </head>
   <header class="header-public">
     <nav class="navbar mt-0 navigation navbar-expand-sm">
       <div class="awesome-book-list">
         <span class="header">Awesome books</span>
-        <button
-          class="btn-info btn btn-sm"
-          type="button"
-          onclick="showList(this), clickMe(this)"
-        >
+        <button class="btn-info btn btn-sm nav-btn-list" type="button">
           list
         </button>
       </div>
       <div class="add-new-container">
-        <button
-          class="nav-btn-add btn btn-info btn-sm"
-          type="button"
-          onclick="addNew(this)"
-        >
+        <button class="nav-btn-add btn btn-info btn-sm" type="button">
           Add new
         </button>
       </div>
       <div class="add-new-container">
-        <button
-          class="nav-btn btn btn-info btn-sm"
-          type="button"
-          onclick="contact(this)"
-        >
+        <button class="nav-btn-contact btn btn-info btn-sm" type="button">
           Contact
         </button>
       </div>
     </nav>
   </header>
-  <body class="bdy" onload="displayDatas(this)">
+  <body class="bdy">
     <main class="wraper">
-      <div class="dispay-data"></div>
+      <div class="display-data"></div>
       <div class="container">
-        <h1 class="page-header">All Awesome Books</h1>
+        <h1 class="page-header">Add book</h1>
         <div id="bookList" class="border"></div>
         <div class="mb-4 textfield-and-btn">
           <div class="textfield-and-bt">
@@ -87,6 +75,6 @@
       <footer>&copy; Najibullah Jafari 2023</footer>
     </main>
 
-    <script src="script.js"></script>
+    <script type="module" src="index.js"></script>
   </body>
 </html>
